% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract Model Coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}, usually the result of a call to \code{linreg()}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
A named numeric vector containing the regression coefficients.
}
\description{
A method for the generic function \code{coef} to extract model coefficients
from an object of class "linreg".
}
\examples{
model <- linreg(Petal.Length ~ Sepal.Length, data = iris)
coef(model)
}
