% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Perform Linear Regression using Ordinary Least Squares}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A `formula` object specifying the model, for example, `Petal.Length ~ Sepal.Length + Species`.}

\item{data}{A `data.frame` containing the variables specified in the formula. The function
includes a pre-processing step to automatically convert numeric-like character or
factor columns into a numeric type.}
}
\value{
An object of S3 class `linreg`. This object is a list containing the following components:
\item{coefficients}{A matrix of the estimated regression coefficients (`beta_hat`).}
\item{fitted.values}{A matrix of the model's fitted values (`y_hat`).}
\item{residuals}{A matrix of the model's residuals (`e_hat`).}
\item{df}{An integer representing the degrees of freedom.}
\item{residual_variance}{A numeric scalar of the residual variance (`v_hat`).}
\item{variance_coefficients}{The variance-covariance matrix of the regression coefficients (`v_beta_hat`).}
\item{t_values}{A matrix of the t-values for each coefficient (`t_beta`).}
\item{p_values}{A matrix of the p-values for each coefficient (`p_beta`).}
}
\description{
This function fits a multiple linear regression model using Ordinary Least Squares (OLS)
based on a provided formula and data frame. It computes regression coefficients,
fitted values, residuals, and key statistical metrics for inference.
}
\examples{
# Fit a linear model on the iris dataset
linreg_model <- linreg(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
}
