% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Extract Model Predictions}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class `linreg`, usually the result of a call to `linreg()`.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
A named numeric vector containing the predicted values (ŷ).
}
\description{
This is a method for the generic function \code{pred} to extract model
predictions from an object of class "linreg". It returns the fitted values from the regression.
}
\examples{
# Create a linreg model
model <- linreg(Petal.Length ~ Sepal.Length, data = iris)
# Extract the predicted values
pred(model)
}
