% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linreg.R
\name{pred}
\alias{pred}
\title{Generic function for model predictions
This is a generic function to extract predictions from various model objects.}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The predictions from the model.
}
\description{
Generic function for model predictions
This is a generic function to extract predictions from various model objects.
}
