% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg objects}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class `linreg`, usually the result of a call to `linreg()`.}

\item{...}{further arguments passed to or from other methods (ignored in this method).}
}
\value{
The original `linreg` object (returned invisibly).
}
\description{
Prints coefficients of a `linreg` model object.
}
\examples{
model <- linreg(formula = Petal.Length ~ Sepal.Length + Sepal.Width, data = iris)
print(model)
}
