% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linreg.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Extract Model Residuals}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class `linreg`, usually the result of a call to `linreg()`.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
A named numeric vector containing the residual values (y - ŷ).
}
\description{
Extracts the vector of residuals from a model object of class "linreg".
}
\examples{
# Create a linreg model
model <- linreg(Petal.Length ~ Sepal.Length, data = iris)
# Extract the residuals
residuals(model)
}
